move location of cmd
